fix some bugs

+++ utils/remove-file.js
@@ -1,12 +1,14 @@
 const fs = require("fs");
 
 const removeFile = (path) => {
-  fs.unlink(path, (err) => {
-    if (err) {
-      console.error("Gagal menghapus file:", err);
-      throwError("Gagal menghapus file", 500);
-    }
-  });
+  if (path) {
+    fs.unlink(path, (err) => {
+      if (err) {
+        console.error("Gagal menghapus file:", err);
+        throwError("Gagal menghapus file", 500);
+      }
+    });
+  }
 };
 
 module.exports = removeFile;
